Update Python and okdata-sdk requirements
Add support for Python 3.13 and drop support for Python 3.8, which has
reached end of life, and Python 3.9 which is very soon reaching end of
life. Python 3.9 is dropped a bit prematurely because it's
incompatible when combining boto3 with urllib3 v2.x.

Also bump the okdata-sdk version requirements to get rid of the
vulnerable urllib3 v1.x dependency.